+++ src/bezzanlabs/capsule/proto.py
@@ -2,6 +2,7 @@
 Prototypes for model classes and supporting functions.
 """
 import typing as tp
+
 from .types import Inputs, Predictions
 
 
@@ -29,6 +30,7 @@ class Regressor(tp.Protocol):
     """
     Defines the base behavior for a Regressor object.
     """
+
     def predict(self, X: Inputs) -> Predictions:
         """
         Defines predict method for a Regressor.