+++ CHANGELOG.md
@@ -3,4 +3,6 @@
 Include your changelog here for each version being deployed
 
 ## v0.0.1
-
+- Adding support for AES-256 encryption of models
+- Returning predictions and predict_proba
+-